visualisation/requirements.txt
pandas~=2.2.3
numpy~=2.1.3
pytz~=2024.2
python-dateutil~=2.9.0.post0
matplotlib~=3.9.4
functionCacher~=1.0.5.2